+++ game_project/entities/characters/player.cpp
@@ -51,8 +51,8 @@ void Player::checkInputs(std::map<int, bool> keyboardArray)
 
 void Player::resize()
 {
-	cardPointsNumber.position = sf::Vector2f(windowInfo.activeSceneWidthPixels + 4 * 32, windowInfo.getWindowHeight() - 32);
-	healthNumber.position = sf::Vector2f(windowInfo.activeSceneWidthPixels + 4 * 32, windowInfo.getWindowHeight() - 64);
+	cardPointsNumber.position = sf::Vector2f((float)(windowInfo.activeSceneWidthPixels + 4 * 32), (float)(windowInfo.getWindowHeight() - 32));
+	healthNumber.position = sf::Vector2f((float)(windowInfo.activeSceneWidthPixels + 4 * 32), (float)(windowInfo.getWindowHeight() - 64));
 
 	cardsInHand.resize();
 }
